requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
sqlalchemy
pyodbc
python-jose
passlib[bcrypt]
python-multipart
